fix(pos): store pos_orders.cod as numeric(12,2) instead of float4
REAL maps to float4 in Postgres — roughly 7 significant digits. That is the
column behind the COD figure on RMO Management, and SUM(cod) accumulates in
float4 as well, so both individual values and totals could drift off the
centavo. Money wants an exact decimal type.

Applied to the live database and mirrored in schema.pg.sql for fresh installs,
plus an idempotent boot migration that converts any other Postgres environment
(it checks udt_name first, so re-running is a no-op). SQLite is untouched: its
REAL is already a 64-bit double and it has no fixed-scale numeric type.

pg returns numeric as a string, so the three API handlers that passed row.cod
straight into JSON now wrap it in Number() — otherwise the response contract
would have quietly changed from 398 to "398.00". Every SUM(cod) consumer
already went through Number() and needed no change.

+++ backend/db/schema.pg.sql
@@ -326,7 +326,7 @@ CREATE TABLE IF NOT EXISTS pos_orders (
   customer_email TEXT,
   page_id TEXT,
   shipping_fee REAL,
-  cod REAL,
+  cod NUMERIC(12,2),
   cash REAL,
   total_discount REAL,
   note TEXT,
